requirements-dashboard.txt
# NBA Simulator - Monitoring Dashboard Requirements
# Install with: pip install -r requirements-dashboard.txt

# Core Web Framework
Flask==3.0.0
flask-cors==4.0.0

# AWS Integration
boto3==1.34.0
botocore==1.34.0

# Database
psycopg2-binary==2.9.9

# Async Support
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Optional Production
# gunicorn==21.2.0
# gevent==23.9.1
# redis==5.0.1
